workers/python/pyproject.toml
[project]
name = "omes"
version = "0.1.0"
description = "Temporal load generator"
authors = [{ name = "Temporal Technologies Inc", email = "sdk@temporal.io" }]
requires-python = "~=3.10"
license = "MIT"
dependencies = [
    "temporalio>=1.23.0,<2",
    "grpcio>=1.60.0,<2",
    "prometheus-client>=0.16.0,<0.17",
    "python-json-logger>=2.0.7,<3",
]

[dependency-groups]
dev = [
    "mypy>=0.961,<0.962",
    "black>=22.3.0,<23",
    "isort>=5.10.1,<6",
]

[tool.hatch.build.targets.sdist]
include = ["**/*.py"]

[tool.hatch.build.targets.wheel]
include = ["**/*.py"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.poe.tasks]
format = [{cmd = "black ."}, {cmd = "isort ."}]
lint = [
  {cmd = "black --check ."},
  {cmd = "isort --check-only ."},
  {ref = "lint-types"},
]
lint-types = "mypy --explicit-package-bases --namespace-packages --exclude 'harness/api/' ."
test = "python -m unittest discover -s tests -p 'test_*.py'"

[tool.isort]
profile = "black"
skip_gitignore = true
skip_glob = ["protos/*", "harness/api/*"]

[tool.black]
extend-exclude = '''
(
  protos/
  | harness/api/
)
'''
